# Try moving contig sorting into sqlite