--- a/spacegraphcats/cdbg/bcalm_to_gxt2.py
+++ b/spacegraphcats/cdbg/bcalm_to_gxt2.py
@@ -318,6 +318,8 @@ def main(argv):
 
     sequences.close()
 
+    return 0
+
 
 if __name__ == "__main__":
     main(sys.argv[1:])