--- a/spacegraphcats/utils/make_bgzf.py
+++ b/spacegraphcats/utils/make_bgzf.py
@@ -41,6 +41,8 @@ def main(argv=sys.argv[1:]):
 
     print("done!")
 
+    return 0
+
 
 if __name__ == "__main__":
     sys.exit(main())