… data/bcalm-BROKEN.dory.k21.unitigs.fa
… data/bcalm.dory.k21.unitigs.fa
… environment.yml
… spacegraphcats/conf/Snakefile
… spacegraphcats/conf/twofoo-short.yaml
… spacegraphcats/search/decompose_catlas.py
… spacegraphcats/search/estimate_query_abundance.py
… spacegraphcats/search/extract_nodes_by_shadow_ratio.py
… spacegraphcats/search/extract_unassembled_nodes.py
… spacegraphcats/search/index_cdbg_by_multifasta.py
… spacegraphcats/search/query_by_hashval.py
… spacegraphcats/search/query_multifasta_by_sig.py
… spacegraphcats/search/test_workflow.py
… spacegraphcats/test_snakemake.py
… spacegraphcats/test_twofoo_short.py