This is an attempt to store less in memory and move more to fast on-disk operations.

So far, this PR:
* puts all the sequences into a sqlite database; does sorting and remapping of IDs there
* moves length and abundance information into the sqlite database, too
* renames the cDBG directory to `{catlas_base}_{ksize}`, e.g. `twofoo_k31`; renames files to no longer contain `{catlas_base}`
* misc cleanup for sourmash 4.0: use hashes, str(name), etc.

In the future, we could explore:
* storing the `neighbors` structure in the SQLite database
* no longer outputting contigs.fa.gz and just leaving the contigs in the sqlite database
* doing pendant removal iteratively (not sure this is possible!)

TODO:
- [x] write test for link structure check
- [x] clean up comments, etc.