@ctb 2020-11-05
## benchmarks against hu-s1

Compare with https://github.com/spacegraphcats/spacegraphcats/pull/310#issuecomment-720096022:

old branch total (everything in memory):
* stable/v1.4: 20m:23s, in 19.6 GB

sort:

* #310: 30m:21s, in 20.3 GB
* this branch: 35m:33s, in 12.9 GB

output gxt:
* #310: 30m:53s, in 17.2 GB 
* this branch: 18m:26s, in 14.9 GB

So, we get some reasonable memory improvement - about 25% - on this data set!

Speed not awesome, but we could probably help fix that by not actually outputting `contigs.fa.gz`, which would also presumably have benefits down the line for search, etc.